HardFaultTrampoline now passes &mut ExceptionFrame and SP
In order to allow for a user provided HardFault handler to modify the
exception frame (e.g. to return to a different PC) the signature for the
function to be used with the `#[HardFault]` handler now takes a `&mut
ExceptionFrame` instead of just a reference, as well as the SP at the
moment of exception entry in order to be able to pop the EXC_RETURN
value.